pyVerein/requirements_development.txt
-r requirements_base.txt
pylint>=1.8.4,<1.9
django-debug-toolbar>=1.9.1,<2.0